requirements.txt
Django==5.2.4
django-admin-sortable2==2.2.8
django-phonenumber-field==8.1.0
phonenumbers==9.0.10
pillow==11.3.0
django-admin-sortable2==2.2.8
qrcode==8.2
python-dotenv==1.1.1